requirements.txt
numpy==1.7.1
Pillow==2.1.0
requests==1.2.3
requests-oauthlib==0.3.2
